update drugban tests and rename `embed.video` APIs
