+++ kale/embed/video/se_res3d.py
@@ -6,7 +6,7 @@
 
 from torch.hub import load_state_dict_from_url
 
-from kale.embed.video.video_res3d import (
+from kale.embed.video.res3d import (
     BasicBlock,
     BasicFLowStem,
     BasicStem,
@@ -17,7 +17,7 @@ from kale.embed.video.video_res3d import (
     R2Plus1dStem,
     VideoResNet,
 )
-from kale.embed.video.video_selayer import get_selayer, SELayerC, SELayerT
+from kale.embed.video.selayer import get_selayer, SELayerC, SELayerT
 
 model_urls = {
     "r3d_18": "https://download.pytorch.org/models/r3d_18-b3b3357e.pth",
